Movement to multiple files and some cleaning house
