+++ testLevels.json
@@ -2,7 +2,7 @@
     "L0": [
         {
             "ts": [
-                "grey.jpeg"
+                "sampleTexture.jpeg"
             ],
             "cS": null,
             "mf": []
